requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
torch==2.1.1
torchvision==0.16.1
pillow==10.1.0
pymongo==4.6.0
weaviate-client==4.1.2
python-dotenv==1.0.0
numpy==1.24.3
pytest==7.4.3
git+https://github.com/openai/CLIP.git
